Finished the set empty, update, and collision
The set(key, value) function adds a new node to the
hashmap using the hash function to obtain the
index within the array's capacity/length. Accepts
two arguments, the key to hash, and the value
of the node itself. I made modifications to the
nodes.js and linkedlist.js to include a
nodeKey property to keep track of what key the
node had. If the hash code index location in
the array is empty, then we will make a new
linked list at that location and append our
node with the key and value in the node itself.
If there is already a linked list at the
hash code index location, then we will check if
the key already exists in any of the nodes in that
linked list. If it does exist, then update the
value with the new value given. If it doesn't
exist, then that means that a different key
was hashed to the same index location, so we would
have to recognize that as a collision and append
it to the end of our linked list bucket.