Finished growth logic and added rehash function
So if the hashmap detects that it is time to
grow based on the product of the load factor
and the current capacity of the array, then it
will double the array capacity and perform
a rehash of EVERY node in there to evenly
distribute the nodes in the new length array.
We always have to rehash after growing because
the capacity changes the hash code we get,
so if we don't rehash the entire hashmap then
we would run the risk of multiple collisions
with new nodes. If we rehash, we can ensure
that the nodes that already existed get put
in their right places based on the new capacity
hash code index calculation, and it would be more
evenly spread out in the array.